refactor: introduce directory features
