+++ src/app/App/MainScreen/NotesList/index.tsx
@@ -1,8 +1,8 @@
 import React, { FC } from 'react';
 import { cn } from '@bem-react/classname';
 
-import { INotesController } from '../../../../core/storage/controllers/notes';
-import { INote, NoteId } from '../../../../types/notes';
+import { INote, NoteId } from '../../../../core/features/notes';
+import { INotesController } from '../../../../core/features/notes/controller';
 import { Stack } from '../../../components/Stack/Stack';
 
 import { getNoteTitle } from '../..';
