+++ src/app/App/NoteEditor/index.tsx
@@ -26,15 +26,15 @@ import { useStore } from 'effector-react';
 import { debounce } from 'lodash';
 import { cn } from '@bem-react/classname';
 
-import { findLinksInText, getResourceIdInUrl } from '../../../core/links';
+import { findLinksInText, getResourceIdInUrl } from '../../../core/features/links';
+import { INote, INoteContent } from '../../../core/features/notes';
+import { ITag } from '../../../core/features/tags/TagsController';
 import {
 	$tags,
 	setActiveTag,
 	tagAttachmentsChanged,
 	tagsChanged,
 } from '../../../core/state/tags';
-import { ITag } from '../../../core/storage/controllers/tags/TagsController';
-import { INote, INoteContent } from '../../../types/notes';
 import { Icon } from '../../components/Icon/Icon.bundle/common';
 import { Stack } from '../../components/Stack/Stack';
 import { SuggestedTagsList } from '../../components/SuggestedTagsList';
