+++ src/app/App/NoteScreen/index.tsx
@@ -14,8 +14,8 @@ import remarkBreaks from 'remark-breaks';
 import remarkGfm from 'remark-gfm';
 import { cn } from '@bem-react/classname';
 
-import { getResourceIdInUrl } from '../../../core/links';
-import { INote } from '../../../types/notes';
+import { getResourceIdInUrl } from '../../../core/features/links';
+import { INote } from '../../../core/features/notes';
 
 import { useAttachmentsRegistry, useFilesRegistry } from '../Providers';
 import {
