+++ src/app/App/Providers.tsx
@@ -1,9 +1,9 @@
 import React, { createContext, FC, useContext } from 'react';
 
-import { AttachmentsController } from '../../core/storage/controllers/attachments/AttachmentsController';
-import { FilesController } from '../../core/storage/controllers/files/FilesController';
-import { INotesController } from '../../core/storage/controllers/notes';
-import { TagsController } from '../../core/storage/controllers/tags/TagsController';
+import { AttachmentsController } from '../../core/features/attachments/AttachmentsController';
+import { FilesController } from '../../core/features/files/FilesController';
+import { INotesController } from '../../core/features/notes/controller';
+import { TagsController } from '../../core/features/tags/TagsController';
 import { SQLiteDatabase } from '../../core/storage/database/SQLiteDatabase/SQLiteDatabase';
 
 export type FileId = string;
