+++ src/core/features/attachments/AttachmentsController.test.ts
@@ -1,7 +1,7 @@
 import { tmpdir } from 'os';
 import { tmpNameSync } from 'tmp';
 
-import { openDatabase } from '../../database/SQLiteDatabase/SQLiteDatabase';
+import { openDatabase } from '../../storage/database/SQLiteDatabase/SQLiteDatabase';
 
 import { AttachmentsController } from './AttachmentsController';
