+++ src/core/features/notes/controller/NotesController.performance.test.ts
@@ -1,7 +1,7 @@
 import { tmpdir } from 'os';
 import { tmpNameSync } from 'tmp';
 
-import { openDatabase } from '../../database/SQLiteDatabase/SQLiteDatabase';
+import { openDatabase } from '../../../storage/database/SQLiteDatabase/SQLiteDatabase';
 
 import { rm } from 'fs/promises';
 import { NotesController } from './NotesController';
