+++ src/core/features/notes/index.ts
